Add support for building as a dependency of std.
Add an optional "rustc-dep-of-std" feature which, when enabled, enables
support for the special dependencies that allow this library to be used
as a dependency of std.

Very few users will want to use this, however I am experimenting with
some changes in [mustang] where having this feature in scopeguard would
be useful.

This also includes a minor warning fix that comes up in the std build.

[mustang]: https://github.com/sunfishcode/mustang